requirements.txt
# Core LLM & API
openai>=1.0.0
tavily-python>=0.3.0

# Web Scraping & Content Extraction
beautifulsoup4>=4.12.0
requests>=2.31.0
trafilatura>=1.6.0
lxml>=4.9.0
html5lib>=1.1

# Database & State Management
sqlalchemy>=2.0.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Token Management
tiktoken>=0.5.0

# Progress & CLI
rich>=13.0.0
typer>=0.9.0

# Export Formats
markdown>=3.5.0
jinja2>=3.1.0
markupsafe>=2.1.0
weasyprint>=60.0

# Utilities
tenacity>=8.2.0
python-dateutil>=2.8.0

# Web Dashboard
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# MCP Server
mcp>=1.0,<2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
